fix(metrics-layer): Change how interval calculation is handled (#66047)

+++ src/sentry/snuba/metrics/query.py
@@ -385,7 +385,9 @@ class MetricsQuery(MetricsQueryValidationRunner):
             interval = self.interval
 
         if self.start and self.end and self.include_series:
-            if (self.end - self.start).total_seconds() / interval > MAX_POINTS:
+            # For this calculation, we decided to round down to the integer since if we get 10.000,x we prefer to allow
+            # the query and lose some data points. On the other hand, if we get 11.000,x we will not allow the query.
+            if int((self.end - self.start).total_seconds() / interval) > MAX_POINTS:
                 raise InvalidParams(
                     "Your interval and date range would create too many results. "
                     "Use a larger interval, or a smaller date range."
